refactor: command, query 패키지 구조 분리
