+++ mohaeng/src/main/java/com/mohaeng/club/club/application/service/command/DelegatePresident.java
@@ -1,6 +1,6 @@
 package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.DelegatePresidentUseCase;
+import com.mohaeng.club.club.application.usecase.command.DelegatePresidentUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
 import com.mohaeng.club.club.exception.ClubException;
