+++ mohaeng/src/main/java/com/mohaeng/club/club/application/service/command/DeleteClub.java
@@ -1,6 +1,6 @@
 package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.DeleteClubUseCase;
+import com.mohaeng.club.club.application.usecase.command.DeleteClubUseCase;
 import com.mohaeng.club.club.domain.event.DeleteClubEvent;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.model.Participant;
