+++ mohaeng/src/main/java/com/mohaeng/club/club/application/usecase/command/ChangeClubRoleNameUseCase.java
@@ -1,4 +1,4 @@
-package com.mohaeng.club.club.application.usecase;
+package com.mohaeng.club.club.application.usecase.command;
 
 /**
  * 모임 역할 이름 변경
