+++ mohaeng/src/main/java/com/mohaeng/club/club/application/usecase/command/CreateClubUseCase.java
@@ -1,4 +1,4 @@
-package com.mohaeng.club.club.application.usecase;
+package com.mohaeng.club.club.application.usecase.command;
 
 public interface CreateClubUseCase {
