+++ mohaeng/src/main/java/com/mohaeng/club/club/presentation/command/DeleteClubRoleController.java
@@ -1,7 +1,7 @@
-package com.mohaeng.club.club.presentation;
+package com.mohaeng.club.club.presentation.command;
 
 import com.mohaeng.authentication.presentation.argumentresolver.Auth;
-import com.mohaeng.club.club.application.usecase.DeleteClubRoleUseCase;
+import com.mohaeng.club.club.application.usecase.command.DeleteClubRoleUseCase;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
