+++ mohaeng/src/test/java/com/mohaeng/club/club/application/service/command/ChangeDefaultRoleTest.java
@@ -1,6 +1,6 @@
 package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.ChangeDefaultRoleUseCase;
+import com.mohaeng.club.club.application.usecase.command.ChangeDefaultRoleUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.model.ClubRole;
 import com.mohaeng.club.club.domain.model.ClubRoleCategory;
