+++ mohaeng/src/test/java/com/mohaeng/club/club/application/service/command/CreateClubRoleTest.java
@@ -1,6 +1,6 @@
 package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.usecase.CreateClubRoleUseCase;
+import com.mohaeng.club.club.application.usecase.command.CreateClubRoleUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.model.ClubRole;
 import com.mohaeng.club.club.domain.model.ClubRoleCategory;
