+++ mohaeng/src/test/java/com/mohaeng/club/club/application/service/command/CreateClubTest.java
@@ -1,7 +1,6 @@
 package com.mohaeng.club.club.application.service.command;
 
-import com.mohaeng.club.club.application.service.command.CreateClub;
-import com.mohaeng.club.club.application.usecase.CreateClubUseCase;
+import com.mohaeng.club.club.application.usecase.command.CreateClubUseCase;
 import com.mohaeng.club.club.domain.model.Club;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
 import com.mohaeng.common.annotation.ApplicationTest;
