+++ mohaeng/src/test/java/com/mohaeng/club/club/presentation/command/DelegatePresidentControllerTest.java
@@ -1,6 +1,6 @@
-package com.mohaeng.club.club.presentation;
+package com.mohaeng.club.club.presentation.command;
 
-import com.mohaeng.club.club.application.usecase.DelegatePresidentUseCase;
+import com.mohaeng.club.club.application.usecase.command.DelegatePresidentUseCase;
 import com.mohaeng.club.club.exception.ClubException;
 import com.mohaeng.club.club.exception.ParticipantException;
 import com.mohaeng.common.ControllerTest;
@@ -17,7 +17,7 @@ import org.springframework.test.web.servlet.ResultActions;
 import static com.mohaeng.club.club.exception.ClubExceptionType.NOT_FOUND_CLUB;
 import static com.mohaeng.club.club.exception.ParticipantExceptionType.NOT_FOUND_PARTICIPANT;
 import static com.mohaeng.club.club.exception.ParticipantExceptionType.NO_AUTHORITY_DELEGATE_PRESIDENT;
-import static com.mohaeng.club.club.presentation.DelegatePresidentController.DELEGATE_PRESIDENT_URL;
+import static com.mohaeng.club.club.presentation.command.DelegatePresidentController.DELEGATE_PRESIDENT_URL;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentRequest;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentResponse;
 import static com.mohaeng.common.fixtures.AuthenticationFixture.BEARER_ACCESS_TOKEN;
