+++ mohaeng/src/test/java/com/mohaeng/club/club/presentation/command/ExpelParticipantControllerTest.java
@@ -1,65 +1,72 @@
-package com.mohaeng.club.club.presentation;
+package com.mohaeng.club.club.presentation.command;
 
-import com.mohaeng.club.club.application.usecase.DeleteClubUseCase;
-import com.mohaeng.club.club.exception.ClubException;
+import com.mohaeng.club.club.application.usecase.command.ExpelParticipantUseCase;
+import com.mohaeng.club.club.exception.ParticipantException;
 import com.mohaeng.common.ControllerTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.test.web.servlet.ResultActions;
 
-import static com.mohaeng.club.club.exception.ClubExceptionType.NOT_FOUND_CLUB;
-import static com.mohaeng.club.club.exception.ClubExceptionType.NO_AUTHORITY_DELETE_CLUB;
-import static com.mohaeng.club.club.presentation.DeleteClubController.DELETE_CLUB_URL;
+import static com.mohaeng.club.club.exception.ParticipantExceptionType.NOT_FOUND_PARTICIPANT;
+import static com.mohaeng.club.club.exception.ParticipantExceptionType.NO_AUTHORITY_EXPEL_PARTICIPANT;
+import static com.mohaeng.club.club.presentation.command.ExpelParticipantController.EXPEL_PARTICIPANT_URL;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentRequest;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentResponse;
 import static com.mohaeng.common.fixtures.AuthenticationFixture.BEARER_ACCESS_TOKEN;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.*;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@DisplayName("DeleteClubController 는 ")
-@WebMvcTest(DeleteClubController.class)
-class DeleteClubControllerTest extends ControllerTest {
+@DisplayName("ExpelParticipantController 는")
+@WebMvcTest(ExpelParticipantController.class)
+class ExpelParticipantControllerTest extends ControllerTest {
 
     @MockBean
-    private DeleteClubUseCase deleteClubUseCase;
+    private ExpelParticipantUseCase expelParticipantUseCase;
+
+    private final Long participantId = 1L;
+    private final Long memberId = 1L;
+    private final Long clubId = 1L;
 
     @Test
-    void 모임의_회장이_모임_제거_요청을_보낸_경우_모임을_제거하고_200을_반환한다() throws Exception {
+    void 모임에서_추방_성공_시_200을_반환한다() throws Exception {
         // given
-        final Long memberId = 1L;
+        doNothing().when(expelParticipantUseCase).command(any());
         setAuthentication(memberId);
 
-        // when & then
+        // when
         ResultActions resultActions = mockMvc.perform(
-                delete(DELETE_CLUB_URL, 1L)
+                delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
                         .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
         ).andExpect(status().isOk());
 
-        BDDMockito.then(deleteClubUseCase).should().command(any());
+        // then
+        verify(expelParticipantUseCase, times(1)).command(any());
+
         resultActions.andDo(
-                document("delete-club",
+                document("expel-participant-from-club",
                         getDocumentRequest(),
                         getDocumentResponse(),
                         requestHeaders(
                                 headerWithName(HttpHeaders.AUTHORIZATION).description("Access Token")
                         ),
                         pathParameters(
-                                parameterWithName("clubId").description("제거할 모임의 ID")
+                                parameterWithName("clubId").description("추방시킬 참여자가 존재하는 모임"),
+                                parameterWithName("participantId").description("추방시킬 참여자(Participant) ID")
                         )
                 )
         );
@@ -68,42 +75,43 @@ class DeleteClubControllerTest extends ControllerTest {
     @Test
     void 인증되지_않은_사용자의_경우_401을_반환한다() throws Exception {
         // given
-        final Long memberId = 1L;
+        doNothing().when(expelParticipantUseCase).command(any());
         setAuthentication(memberId);
 
-        // when & then
         ResultActions resultActions = mockMvc.perform(
-                        delete(DELETE_CLUB_URL, 1L)
+                        delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
                 )
+                .andDo(print())
                 .andExpect(status().isUnauthorized());
 
-        BDDMockito.then(deleteClubUseCase).shouldHaveNoInteractions();
+        // when & then
+        verify(expelParticipantUseCase, times(0)).command(any());
 
         resultActions.andDo(
-                document("delete-club(No Access Token)",
+                document("expel-participant-from-club(No Access Token)",
                         getDocumentResponse()
                 )
         );
     }
 
     @Test
-    void 모임_ID_가_없는_경우_404를_반환한다() throws Exception {
+    void 참가자를_찾을수_없는_경우_404를_반환한다() throws Exception {
         // given
-        final Long memberId = 1L;
+        doThrow(new ParticipantException(NOT_FOUND_PARTICIPANT))
+                .when(expelParticipantUseCase).command(any());
+
         setAuthentication(memberId);
-        doThrow(new ClubException(NOT_FOUND_CLUB))
-                .when(deleteClubUseCase).command(any());
+
+        ResultActions resultActions = mockMvc.perform(
+                delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
+                        .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
+        ).andExpect(status().isNotFound());
 
         // when & then
-        ResultActions resultActions = mockMvc.perform(
-                        delete(DELETE_CLUB_URL, 1L)
-                                .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
-                )
-                .andExpect(status().isNotFound());
+        verify(expelParticipantUseCase, times(1)).command(any());
 
-        BDDMockito.then(deleteClubUseCase).should().command(any());
         resultActions.andDo(
-                document("delete-club(Nonexistent Club ID)",
+                document("expel-participant-from-club(Nonexistent Participant)",
                         getDocumentResponse()
                 )
         );
@@ -112,22 +120,22 @@ class DeleteClubControllerTest extends ControllerTest {
     @Test
     void 요청자가_회장이_아닌_경우_403을_반환한다() throws Exception {
         // given
-        final Long memberId = 1L;
-        doThrow(new ClubException(NO_AUTHORITY_DELETE_CLUB))
-                .when(deleteClubUseCase).command(any());
+        doThrow(new ParticipantException(NO_AUTHORITY_EXPEL_PARTICIPANT))
+                .when(expelParticipantUseCase).command(any());
 
         setAuthentication(memberId);
 
         ResultActions resultActions = mockMvc.perform(
-                delete(DELETE_CLUB_URL, 1L)
-                        .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
-        ).andExpect(status().isForbidden());
+                        delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
+                                .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
+                ).andDo(print())
+                .andExpect(status().isForbidden());
 
         // when & then
-        BDDMockito.then(deleteClubUseCase).should().command(any());
+        verify(expelParticipantUseCase, times(1)).command(any());
 
         resultActions.andDo(
-                document("delete-club(requester does not president)",
+                document("expel-participant-from-club(requester does not president)",
                         getDocumentResponse()
                 )
         );
