+++ mohaeng/src/test/java/com/mohaeng/club/club/presentation/command/LeaveClubControllerTest.java
@@ -1,6 +1,7 @@
-package com.mohaeng.club.club.presentation;
+package com.mohaeng.club.club.presentation.command;
 
-import com.mohaeng.club.club.application.usecase.ExpelParticipantUseCase;
+import com.mohaeng.club.club.application.usecase.command.LeaveClubUseCase;
+import com.mohaeng.club.club.exception.ClubException;
 import com.mohaeng.club.club.exception.ParticipantException;
 import com.mohaeng.common.ControllerTest;
 import org.junit.jupiter.api.DisplayName;
@@ -12,14 +13,16 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.test.web.servlet.ResultActions;
 
-import static com.mohaeng.club.club.exception.ParticipantExceptionType.NOT_FOUND_PARTICIPANT;
-import static com.mohaeng.club.club.exception.ParticipantExceptionType.NO_AUTHORITY_EXPEL_PARTICIPANT;
-import static com.mohaeng.club.club.presentation.ExpelParticipantController.EXPEL_PARTICIPANT_URL;
+import static com.mohaeng.club.club.exception.ClubExceptionType.NOT_FOUND_CLUB;
+import static com.mohaeng.club.club.exception.ParticipantExceptionType.PRESIDENT_CAN_NOT_LEAVE_CLUB;
+import static com.mohaeng.club.club.presentation.command.LeaveClubController.LEAVE_CLUB_URL;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentRequest;
 import static com.mohaeng.common.ApiDocumentUtils.getDocumentResponse;
 import static com.mohaeng.common.fixtures.AuthenticationFixture.BEARER_ACCESS_TOKEN;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.BDDMockito.*;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -31,42 +34,40 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@DisplayName("ExpelParticipantController 는")
-@WebMvcTest(ExpelParticipantController.class)
-class ExpelParticipantControllerTest extends ControllerTest {
+@DisplayName("LeaveClubController 은")
+@WebMvcTest(LeaveClubController.class)
+class LeaveClubControllerTest extends ControllerTest {
 
     @MockBean
-    private ExpelParticipantUseCase expelParticipantUseCase;
-
-    private final Long participantId = 1L;
-    private final Long memberId = 1L;
-    private final Long clubId = 1L;
+    private LeaveClubUseCase leaveClubUseCase;
 
     @Test
-    void 모임에서_추방_성공_시_200을_반환한다() throws Exception {
+    void 모임_탈퇴_성공_시_200을_반환한다() throws Exception {
         // given
-        doNothing().when(expelParticipantUseCase).command(any());
+        final Long participantId = 1L;
+        final Long memberId = 1L;
+
+        doNothing().when(leaveClubUseCase).command(any());
         setAuthentication(memberId);
 
         // when
         ResultActions resultActions = mockMvc.perform(
-                delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
+                delete(LEAVE_CLUB_URL, participantId)
                         .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
         ).andExpect(status().isOk());
 
         // then
-        verify(expelParticipantUseCase, times(1)).command(any());
+        then(leaveClubUseCase).should().command(any());
 
         resultActions.andDo(
-                document("expel-participant-from-club",
+                document("leave-club",
                         getDocumentRequest(),
                         getDocumentResponse(),
                         requestHeaders(
                                 headerWithName(HttpHeaders.AUTHORIZATION).description("Access Token")
                         ),
                         pathParameters(
-                                parameterWithName("clubId").description("추방시킬 참여자가 존재하는 모임"),
-                                parameterWithName("participantId").description("추방시킬 참여자(Participant) ID")
+                                parameterWithName("clubId").description("탈퇴하려는 모임 Id")
                         )
                 )
         );
@@ -75,67 +76,74 @@ class ExpelParticipantControllerTest extends ControllerTest {
     @Test
     void 인증되지_않은_사용자의_경우_401을_반환한다() throws Exception {
         // given
-        doNothing().when(expelParticipantUseCase).command(any());
+        final Long participantId = 1L;
+        final Long memberId = 1L;
+        doNothing().when(leaveClubUseCase).command(any());
         setAuthentication(memberId);
 
         ResultActions resultActions = mockMvc.perform(
-                        delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
+                        delete(LEAVE_CLUB_URL, participantId)
                 )
                 .andDo(print())
                 .andExpect(status().isUnauthorized());
 
         // when & then
-        verify(expelParticipantUseCase, times(0)).command(any());
+        then(leaveClubUseCase).shouldHaveNoInteractions();
 
         resultActions.andDo(
-                document("expel-participant-from-club(No Access Token)",
+                document("leave-club(No Access Token)",
                         getDocumentResponse()
                 )
         );
     }
 
     @Test
-    void 참가자를_찾을수_없는_경우_404를_반환한다() throws Exception {
+    void 모임_ID가_없는_경우_404_를_반환한다() throws Exception {
         // given
-        doThrow(new ParticipantException(NOT_FOUND_PARTICIPANT))
-                .when(expelParticipantUseCase).command(any());
+        final Long participantId = 1L;
+        final Long memberId = 1L;
+        doThrow(new ClubException(NOT_FOUND_CLUB))
+                .when(leaveClubUseCase).command(any());
 
         setAuthentication(memberId);
 
         ResultActions resultActions = mockMvc.perform(
-                delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
-                        .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
-        ).andExpect(status().isNotFound());
-
-        // when & then
-        verify(expelParticipantUseCase, times(1)).command(any());
-
-        resultActions.andDo(
-                document("expel-participant-from-club(Nonexistent Participant)",
-                        getDocumentResponse()
-                )
-        );
-    }
-
-    @Test
-    void 요청자가_회장이_아닌_경우_403을_반환한다() throws Exception {
-        // given
-        doThrow(new ParticipantException(NO_AUTHORITY_EXPEL_PARTICIPANT))
-                .when(expelParticipantUseCase).command(any());
-
-        setAuthentication(memberId);
-
-        ResultActions resultActions = mockMvc.perform(
-                        delete(EXPEL_PARTICIPANT_URL, clubId, participantId)
+                        delete(LEAVE_CLUB_URL, participantId)
                                 .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
                 ).andDo(print())
-                .andExpect(status().isForbidden());
+                .andExpect(status().isNotFound());
 
         // when & then
-        verify(expelParticipantUseCase, times(1)).command(any());
+        then(leaveClubUseCase).should().command(any());
 
         resultActions.andDo(
-                document("expel-participant-from-club(requester does not president)",
+                document("leave-club(Nonexistent Club ID)",
+                        getDocumentResponse()
+                )
+        );
+    }
+
+    @Test
+    void 회장이_모임_탈퇴_요청을_한_경우_400_을_반환한다() throws Exception {
+        // given
+        final Long participantId = 1L;
+        final Long memberId = 1L;
+        doThrow(new ParticipantException(PRESIDENT_CAN_NOT_LEAVE_CLUB))
+                .when(leaveClubUseCase).command(any());
+
+        setAuthentication(memberId);
+
+        ResultActions resultActions = mockMvc.perform(
+                        delete(LEAVE_CLUB_URL, participantId)
+                                .header(HttpHeaders.AUTHORIZATION, BEARER_ACCESS_TOKEN)
+                ).andDo(print())
+                .andExpect(status().isBadRequest());
+
+        // when & then
+        verify(leaveClubUseCase, times(1)).command(any());
+
+        resultActions.andDo(
+                document("leave-club(president requests to leave the club)",
                         getDocumentResponse()
                 )
         );
